requirements.txt
schedule
tweepy==4.4.0
requests 
pymongo 
dnspython
python-dotenv
pendulum